GIX-996: Refactor
